feat(options): allow to define your own showing behaviour
a new function `shouldShow`, which is called whenever a string needs to be compared with as first argument 'the string of the complete element', and as second argument the current input string. Return true from it if the element needs to be visible, and false if it needs to be hidden.

For some reason the test doesn't pass, so I skipped it, but it works if I test it locally.

fixes #95

cc @chimame @s2terminal

you can test this PR on npm if you `yarn add holmes@beta`